simplify server & add extentions & update webpack
