+++ client/client.js
@@ -1,3 +1,3 @@
-import Overiew from './app/Overview (Main)/Overview';
+import Overiew from './app/Overview (Main)/Overview.jsx';
 
 window.Overiew = Overiew;
