+++ client/server.js
@@ -1,3 +1,3 @@
-import Overview from './app/Overview (Main)/Overview';
+import Overview from './app/Overview (Main)/Overview.jsx';
 
 export default Overview;
